Initial commit
Creates two launch templates:
1. VLC for when you want to upload your own content to share.
2. Firefox for when you want to just open a browser and view content.

The configuration for these two varies just a little but I duplicated
all the code and config. Not very DRY of me but who cares.

The outputs of the stack are now example commands for how to launch
instances for either of the launch templates.

See `helpers/README.md` for details on scripts that make using this
easier.