Use ASGs, create user with minimal permissions
Creates a new user that has minimal permissions for spinning up and
spinning down instances.

To accomplish this, I moved to using ASGs instead of just calling
`run-instances` with a launch template. `run-instances` is a pretty big
surface area to expose and give creds for, and there are ways to
constrain to just a certain launch template but that doesn't put limits
on how many instances could get launched.

Having an ASG allows for that user to only be able to set
`desired-capacity`. The ASGs can have a max of 1 instance each, and the
ASGs are tagged so that the user can only manipulate `desired-capacity`
on those ASGs.

Scripts in `helpers/` updated to make use of ASGs.